+++ src/main/java/com/upokecenter/numbers/ERational.java
@@ -2121,7 +2121,7 @@ return this.isZero() ? ((int)0) : this.ToEInteger().ToInt32Checked();
  if (!this.isFinite()) {
  throw new ArithmeticException("Value is infinity or NaN");
 }
-return this.isZero() ? ((long)0) : this.ToEInteger().ToInt64Checked();
+return this.isZero() ? 0L : this.ToEInteger().ToInt64Checked();
 }
 
     /**
@@ -2132,7 +2132,7 @@ return this.isZero() ? ((long)0) : this.ToEInteger().ToInt64Checked();
      * value is infinity or not-a-number.
      */
     public long ToInt64Unchecked() {
- return this.isFinite() ? this.ToEInteger().ToInt64Unchecked() : (long)0;
+ return this.isFinite() ? this.ToEInteger().ToInt64Unchecked() : 0L;
 }
 
     /**
@@ -2148,7 +2148,7 @@ return this.isZero() ? ((long)0) : this.ToEInteger().ToInt64Checked();
  if (!this.isFinite()) {
  throw new ArithmeticException("Value is infinity or NaN");
 }
- return this.isZero() ? ((long)0) : this.ToEIntegerIfExact().ToInt64Checked();
+ return this.isZero() ? 0L : this.ToEIntegerIfExact().ToInt64Checked();
 }
 
     /**
